src/app/ui/game/battle-field/battle-field.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { BattleFieldComponent } from './battle-field.component';

describe('Component: BattleField', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [BattleFieldComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([BattleFieldComponent],
      (component: BattleFieldComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(BattleFieldComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(BattleFieldComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-battle-field></app-battle-field>
  `,
  directives: [BattleFieldComponent]
})
class BattleFieldComponentTestController {
}

